+++ CHANGELOG.rst
@@ -12,6 +12,7 @@ Changelog
 * :issue:`34`: Update commands :command:`run-ngs-pipeline` and :command:`run-chip-pipeline` to load large VCF files significantly faster by allowing random access. This also means, from now on, input VCF files must be BGZF compressed (.gz) and indexed (.tbi).
 * :issue:`36`: Update phenotype data for CACNA1S, CFTR, IFNL3, RYR1 (thanks `@NTNguyen13 <https://github.com/NTNguyen13>`__).
 * :pr:`39`: Add new gene F5 (thanks `@NTNguyen13 <https://github.com/NTNguyen13>`__).
+* Update :command:`import-variants` command to be able to subset/exclude specified samples.
 
 0.9.0 (2021-12-05)
 ------------------